+++ plugin-offline/src/main/java/com/mapbox/mapboxsdk/plugins/offline/ui/RegionSelectedCallback.java
@@ -1,0 +1,10 @@
+package com.mapbox.mapboxsdk.plugins.offline.ui;
+
+import com.mapbox.mapboxsdk.offline.OfflineTilePyramidRegionDefinition;
+
+public interface RegionSelectedCallback {
+
+  void onSelected(OfflineTilePyramidRegionDefinition definition, String regionName);
+
+  void onCancel();
+}